requirements.txt
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.1.0
django-debug-toolbar==4.4.6
django-filter~=22.1
django-money==3.5.3
drf-spectacular
Faker~=22.5.1
markdown==3.5.1
markdownify==0.11.6
mysqlclient==2.1.1
mysql-connector-python~=9.0.0
polib~=1.2.0
pydantic==1.10.9
python-dotenv==1.0.1
stripe==5.4.0
requests~=2.31.0
pytz~=2023.3.post1
PyJWT~=2.8.0
cryptography~=42.0.0
drf-yasg==1.21.7
